# apps/backend/auth-service/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "auth-service",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
...
		"/v1/auth/admin/database/createAdmin": {
			"post": {
				"operationId": "CreateAdmin",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a new admin in the database",
				"tags": [
					"Admin Management"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"profile": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"email",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/admin/database/getAllAdmins": {
			"get": {
				"operationId": "GetAllAdmin",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"total",
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get all admins from the database",
				"tags": [
					"Admin Management"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/auth/admin/database/deleteAdmin/{adminId}": {
			"delete": {
				"operationId": "DeleteAdmin",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete an admin from the database",
				"tags": [
					"Admin Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "adminId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/adminOptions/database/getAllUsers": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"total",
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get all users from the database",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/auth/adminOptions/database/getUserById/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get a user by ID from the database",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/adminOptions/database/createUser": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a new user in the database",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"profile": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"email",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/adminOptions/database/updateUser/{id}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Update a user in the database",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"profile": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/adminOptions/database/delete-user/{id}": {
			"post": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete a user from the database",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/adminOptions/database/delete-all-users": {
			"post": {
				"operationId": "DeleteAllUsers",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete all users from the database",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/auth/adminOptions/cognito/listUsersCognito": {
			"get": {
				"operationId": "ListUsersCognito",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"total",
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "List all users from the Cognito",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/auth/adminOptions/cognito/getUsersCognito/{id}": {
			"get": {
				"operationId": "GetUserCognitoByUserName",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get a user from Cognito by username",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/adminOptions/cognito/delete-user/{id}": {
			"post": {
				"operationId": "DeleteUserCognito",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete a user from the Cognito",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/adminOptions/cognito/delete-user/{email}": {
			"delete": {
				"operationId": "CognitoDeleteUserByEmail",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete a user from Cognito by email",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/adminOptions/cognito/delete-all-users": {
			"delete": {
				"operationId": "CognitoDeleteAllUsers",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete all users from Cognito",
				"tags": [
					"Admin Options"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}